requirements.txt
langchain==0.2.7
langchain-ollama
langchain-huggingface
langchain-community
psycopg2-binary
sentence-transformers
torch
faiss-cpu
requests
